Improvements, robustness, permission handling, new features

+++ src/commands/invite/set.js
@@ -1,9 +1,9 @@
-const { EmbedBuilder, SlashCommandBuilder, PermissionsBitField, ChannelType } = require("discord.js");
-const { embedSuccess, embedError } = require("../../utils/colors.js");
+const { EmbedBuilder, SlashCommandBuilder, PermissionsBitField, ChannelType, Colors } = require("discord.js");
 
 module.exports = {
     category: "invites",
     cooldown: 5,
+    botPermissions: [PermissionsBitField.Flags.ManageGuild, PermissionsBitField.Flags.CreateInstantInvite],
     data: new SlashCommandBuilder()
         .setName("invite-set")
         .setDescription("Sets the official guild invitation.")
@@ -16,14 +16,14 @@ module.exports = {
         let invitationChannel = interaction.options.getChannel("channel") || interaction.channel;
 
         if (invitationChannel && invitationChannel.type !== ChannelType.GuildText) {
-            const embed = new EmbedBuilder().setColor(embedError).setDescription("The channel must be a valid text channel.");
+            const embed = new EmbedBuilder().setColor(Colors.Red).setDescription("The channel must be a valid text channel.");
             return interaction.reply({ embeds: [embed] });
         }
 
         if (guildConfig.invitationCode) {
             // TODO: Ask the user for a confirmation, ¿Are you sure? This will delete your current official invitation.
             let invite = await interaction.guild.invites.fetch(guildConfig.invitationCode);
-            await invite.delete();
+            if (invite) await invite.delete();
         }
 
         const invite = await invitationChannel.createInvite({
@@ -34,7 +34,7 @@ module.exports = {
         guildConfig.invitationCode = invite.code;
         await guildConfig.save();
 
-        const embed = new EmbedBuilder().setColor(embedSuccess).setDescription(`The official guild invitation has ben set to ${invite.url}`);
+        const embed = new EmbedBuilder().setColor(Colors.Green).setDescription(`The official guild invitation has ben set to ${invite.url}`);
         await interaction.reply({ embeds: [embed] });
     },
 };
